REFS : #754 made changes to take in text file for observations and images along with mapping files and tested with ifoundbutterflies dataset

+++ grails-app/services/species/AbstractMetadataService.groovy
@@ -73,7 +73,7 @@ class AbstractMetadataService extends AbstractObjectService {
 
     def update(instance, params, klass = null, update=true) {
         if(klass && params.externalId) {
-            instance = klass.findByExternalId(params.externalId.toLong());
+            instance = klass.findByExternalId(params.externalId);
             if(!instance)
                 instance = klass.newInstance();
             else if(!update) {
